add admin access type for retrieving faqs data

+++ app/Http/Controllers/FAQSController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\faqs;
+use Illuminate\Support\Facades\Auth;
 
 class FAQSController extends Controller
 {
@@ -22,6 +23,11 @@ class FAQSController extends Controller
 
     public function store(Request $request)
     {
+
+        if (Auth::user()->type != 'admin') {
+            return response()->json(['message' => 'Permission denied'], 401);
+        }
+
         $faq = new faqs();
         $faq->title = $request->title;
         $faq->definition = $request->definition;
@@ -43,15 +49,27 @@ class FAQSController extends Controller
 
     public function update(Request $request, faqs $faq)
     {
-        $faq->title = $request->input('title');
-        $faq->definition = $request->input('definition');
-        $faq->save();
+        if (Auth::user()->type != 'admin') {
+            return response()->json(['message' => 'Permission denied'], 401);
+        }
 
+        if ($request->filled('title')) {
+            $faq->title = $request->input('title');
+        }
+        if ($request->filled('definition')) {
+            $faq->definition = $request->input('definition');
+        }
+
+        $faq->save();
         return response()->json(['message' => 'Faq updated successfully']);
     }
 
     public function destroy(faqs $faq)
     {
+        if (Auth::user()->type != 'admin') {
+            return response()->json(['message' => 'Permission denied'], 401);
+        }
+
         $faq->delete();
 
         return response()->json(['message' => 'Faq deleted successfully']);
